Decomposition (created cartographedges.js)
