+++ web/iui2017.html
@@ -301,10 +301,8 @@
 
     <script src="js/mingle/kdtree.js"></script>
     <script src="js/mingle/philogl.js"></script>
-    <canvas id="canvas"></canvas><script src="js/mingle/index.js"></script>
     <script src="js/mingle/graph.js"></script>
     <script src="js/mingle/mingle.js"></script>
-     <!--<script src="js/example.js"></script>-->
 
 
     <!-- Main tangram library -->
@@ -313,6 +311,7 @@
 
     <!-- -->
     <script src="js/cartograph.js"></script>
+    <script src="js/cartographedges.js"></script>
 
     <script>
